+++ moldrug/__init__.py
@@ -5,5 +5,7 @@
     Source Code: https://github.com/ale94mleon/moldrug
 """
 from moldrug._version import __version__, __version_tuple__
+__version__
+__version_tuple__
 __author__ = "Alejandro Martínez León"
 __email__ = "ale94mleon@gmail.com"