+++ docs/source/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Changed
+
+- Name of `moldrug.fitness.get_mol_cost` to `moldrug.fitness.__get_mol_cost` function.
+- The class `moldrug.utils.GA` does not have any more the method `roulette_wheel_selection`; now is part a function that could be called from `moldrug.utils`
+
+### Fixed
+
+- Cleaning the code
+
 ## [2.1.12] - 2022.09.29
 
 ### Added